% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_base_rate.R
\name{add_base_rate}
\alias{add_base_rate}
\title{Create data of base application rate information for a single input}
\usage{
add_base_rate(base_input_name, base_unit, base_rate)
}
\arguments{
\item{base_input_name}{(string) input name}

\item{base_unit}{(string) unit of input applied at base}

\item{base_rate}{(numeric) amount of input applied}
}
\value{
data.frame of base rate information
}
\description{
Create data of base application rate information for a single input when applicable. This can be added when preparing trial rates using prepare_rates().
}
\examples{
base_info <- add_base_rate(
  "uan28",
  "gallons",
  15
  )

}
