% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_rates.R
\name{add_blocks}
\alias{add_blocks}
\title{Add blocks to trial design}
\usage{
add_blocks(td)
}
\arguments{
\item{td}{trial design made by applying assign_rates() to experimental plots made by make_exp_plots()}
}
\value{
trial design with block_id added
}
\description{
Delineate blocks on a trial design and assign block id to all the plots
}
\examples{
#--- load rate information ---#
data(trial_design)

#--- add blocks ---#
td_with_blocks <- add_blocks(trial_design)

#--- take a look ---#
td_with_blocks$trial_design

#--- visualize ---#
library(ggplot2)
ggplot(td_with_blocks$trial_design[[1]]) +
  geom_sf(aes(fill = factor(block_id))) +
  geom_sf_text(aes(label = plot_id_within_block)) +
  scale_fill_discrete(name = "block_id") +
  theme_void()

}
