% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_rates.R
\name{change_rate_by_block}
\alias{change_rate_by_block}
\title{Change the assigned rate by block}
\usage{
change_rate_by_block(form, block_ids, block_rates, td)
}
\arguments{
\item{form}{(character) input name}

\item{block_ids}{(numeric) vector of block_ids}

\item{block_rates}{(numeric) vector of rates}

\item{td}{trial design}
}
\value{
trial design with changed rates
}
\description{
Change the assigned rates by block on trial design
}
\examples{
#--- load rate information ---#
data(trial_design)

#--- add blocks ---#
td_with_blocks <- add_blocks(trial_design)

#--- change rates of some blocks ---#
block_ids <- c(3, 5)
block_rates <- c(180, 180)

td_changed <- change_rate_by_block("NH3", block_ids, block_rates, td_with_blocks)

#--- visualize ---# 
library(ggplot2)
ggplot(td_changed$trial_design[[1]]) +
  geom_sf(aes(fill = factor(rate))) +
  scale_fill_viridis_d() +
  theme_void()
}
