% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_rates_manually.R
\name{change_rate_by_plot}
\alias{change_rate_by_plot}
\title{Change the assigned rate by plot}
\usage{
change_rate_by_plot(td, input_name, strip_ids, plot_ids, new_rate)
}
\arguments{
\item{td}{trial design}

\item{input_name}{(character) input name}

\item{strip_ids}{(numeric) vector of strip_ids}

\item{plot_ids}{(numeric) vector of plot_ids}

\item{new_rate}{(numeric) scalar number that is assigned to all the plots specified using strip_ids and plot_ids}
}
\value{
trial design with changed rates
}
\description{
Change the assigned rates by plot on trial design
}
\examples{
#--- load rate information ---#
data(td_single_input)

#--- change rates of some strips ---#
strip_ids <- 1:5
plot_ids <- 5:10
new_rate <- 200

td_modified <- change_rate_by_plot(td_single_input, "NH3", strip_ids, plot_ids, new_rate)

#--- visualize ---#
viz(td_modified)
}
