% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_rates_manually.R
\name{change_rate_by_strip}
\alias{change_rate_by_strip}
\title{Change the assigned rate by strip}
\usage{
change_rate_by_strip(td, input_name, strip_ids, new_rates)
}
\arguments{
\item{td}{trial design}

\item{input_name}{(character) input name}

\item{strip_ids}{(numeric) vector of strip_ids}

\item{new_rates}{(numeric) vector of rates}
}
\value{
trial design with changed rates
}
\description{
Change the assigned rates by strip on trial design
}
\examples{
#--- load rate information ---#
data(td_single_input)

#--- change rates of some strips ---#
strip_ids <- c(1, 6, 11, 16, 21)
new_rates <- rep(0, length(strip_ids))

td_modified <- change_rate_by_strip(td_single_input, "NH3", strip_ids, new_rates)

#--- visualize ---#
viz(td_modified)
}
