% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{check_cor_inputs}
\alias{check_cor_inputs}
\title{Check the correlation of the two inputs}
\usage{
check_cor_inputs(td)
}
\arguments{
\item{td}{trial design with rates assigned created by assign_rates()}
}
\value{
table
}
\description{
Check the correlation between the rates of the two inputs for a two-input experiment.
}
\examples{
#--- load a trial design for a two-input experiment ---#
data(td_two_input)

#--- check correlation ---#
check_cor_inputs(td_two_input)
}
