% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{check_ortho_inputs}
\alias{check_ortho_inputs}
\title{Check the correlation of the two inputs}
\usage{
check_ortho_inputs(td)
}
\arguments{
\item{td}{trial design for a two-input experiment with rates assigned}
}
\value{
table
}
\description{
Check the correlation between the rates of the two inputs for a two-input experiment.
}
\examples{
#--- load a trial design for a two-input experiment ---#
\donttest{
data(td_two_input)

#--- check correlation ---#
check_ortho_inputs(td_two_input)
}
}
