% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_niplorp.R
\name{design_niplorp}
\alias{design_niplorp}
\title{Design a Nitrogen-induced Profit Loss Recovery Program (NIPLORP) experiment}
\usage{
design_niplorp(exp_data, N_ref, N_reduced, ref_pct = 0.25, N_unit)
}
\arguments{
\item{exp_data}{Experiment plots generated using make_exp_plots().}

\item{N_ref}{A numeric value representing the nitrogen rate for the reference plots.}

\item{N_reduced}{A numeric value representing the nitrogen rate for the plots with reducend nitrogen.}

\item{ref_pct}{A numeric value between 0 and 1 representing the proportion of reference plots in the design. Default is 0.25.}

\item{N_unit}{A character string representing the unit of nitrogen rate.}
}
\value{
A tibble containing the trial design with assigned nitrogen rates.
}
\description{
Design an experiment for the Nitrogen-induced Profit Loss Recovery Program (NIPLORP)
}
\examples{

data(exp_data)

td <- design_niplorp(exp_data, N_ref = 200, N_reduced = 120, N_unit = "lb", ref_pct = 0.2)

viz(td)
}
