% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{Diagnose the trial design}
\usage{
diagnose(trial_design, field_data = NA)
}
\arguments{
\item{trial_design}{trial design with rates assigned created by assign_rates()}

\item{field_data}{(list) list of sf}
}
\value{
table
}
\description{
Check the correlation between the input rate and observed characteristics. It checks the correlation between two inputs for a two-input case.
}
