% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_data}
\alias{exp_data}
\title{Experiment data}
\format{
tbl_df tbl data.frame `exp_data`
A data frame with 1 rows and 9 columns:
\describe{
  \item{input_name}{input name}
  \item{harvester_width}{width of the harvester}
  \item{plot_width}{width of the plots to be made}
  \item{field_sf}{field boundary as an sf object}
  \item{headland}{headland as an sf object}
  \item{exp_plots}{experiment plots as an sf object}
  \item{ab_lines}{ab-lines for the applicator/planter as an sf object}
  \item{harvest_ab_lines}{ab-lines for the harvester as an sf object}
  \item{abline_type}{(character) one of "free", "lock", "none" indicating the way ab-line is (or not) created}
}
}
\usage{
exp_data
}
\description{
Data on the experiment created by running the `make_exp_plot()` function, which includes various sf objects (e.g., experiment plots, ab-line, headland, etc). This data exists only for the purpose of making examples in some function references succinct.
}
\keyword{datasets}
