% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_spatial.R
\name{make_ablines}
\alias{make_ablines}
\title{## Get the ab-line}
\usage{
make_ablines(
  ab_sf,
  ab_lines_data,
  base_ab_lines_data,
  plot_width,
  machine_width,
  abline_type
)
}
\description{
## Get the ab-line
}
