% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_exp_plots.R
\name{make_exp_plots}
\alias{make_exp_plots}
\title{Make trial design}
\usage{
make_exp_plots(
  input_plot_info,
  boundary_file,
  harvester_width,
  abline_file = NA,
  abline_type = "free",
  headland_length = NA,
  side_length = NA,
  min_plot_length = 200,
  max_plot_length = 300,
  length_unit = "meter",
  perpendicular = FALSE,
  file_name_append = file_name_append
)
}
\arguments{
\item{input_plot_info}{information of plots created by make_input_plot()}

\item{boundary_file}{(string) name of the field boundary file}

\item{harvester_width}{(numeric) width of the harvester}

\item{abline_file}{(string) name of the ab-line file}

\item{abline_type}{(string) the type of ab-line generation ("free", "lock", "none")}

\item{headland_length}{(numeric, default = NA) length of the headland}

\item{side_length}{(numeric, default = NA) length of the sides}

\item{min_plot_length}{(default = 200 feet) minimum plot length}

\item{max_plot_length}{(default = 300 feet) maximum plot length}

\item{length_unit}{(default = "feet")}

\item{perpendicular}{logical}

\item{file_name_append}{}
}
\description{
Make trial design and return trial design, harvester ab-line, and applicator/planter ab-line.
}
