% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_exp_plots.R
\name{make_input_plot_data}
\alias{make_input_plot_data}
\title{Create a data frame of trial information for a single input}
\usage{
make_input_plot_data(form, unit, plot_width, machine_width, section_num)
}
\arguments{
\item{form}{type of input (e.g., UAN32, seed)}

\item{unit}{unit of length of plot_width and machine_width}

\item{plot_width}{width of plots}

\item{machine_width}{machine width (applicator, planter)}

\item{section_num}{number of sections of the machine}
}
\description{
Create a data frame of trial information for a single input
}
