% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_rates.R
\name{make_input_rate_data}
\alias{make_input_rate_data}
\title{Create data of input rate information for a single input}
\usage{
make_input_rate_data(
  plot_info,
  gc_rate,
  unit,
  rates = NULL,
  min_rate = NULL,
  max_rate = NULL,
  num_rates = 5,
  design_type = NA
)
}
\arguments{
\item{plot_info}{plot information created by make_input_plot_data}

\item{gc_rate}{numeric: rate the grower would have chosen if not running an experiment}

\item{unit}{(string) unit of input}

\item{rates}{(default is NULL)}

\item{min_rate}{numeric minimum input rate}

\item{max_rate}{numeric maximum input rate}

\item{num_rates}{numeric (default is 5)}

\item{design_type}{(string) type of rate design. available options are "jcl", "sparse", and "ejca"}
}
\value{
data.frame of input rate information
}
\description{
Create data of input rate information for a single input. This can be used to assign rates to experimentl plots using assign_rates()
}
