% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_exp_plots.R, R/utility_spatial.R
\name{make_plot_edge_line}
\alias{make_plot_edge_line}
\title{## Get the edge of the experiment for the second input}
\usage{
make_plot_edge_line(
  ab_lines_data,
  create_plot_edge_line,
  base_ab_lines_data,
  plot_width
)

make_plot_edge_line(
  ab_lines_data,
  create_plot_edge_line,
  base_ab_lines_data,
  plot_width
)
}
\description{
## Get the edge of the experiment for the second input

## Get the edge of the experiment for the second input
}
