% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_exp_plots.R
\name{make_trial_plots_by_input}
\alias{make_trial_plots_by_input}
\title{# Create strips
# Shift the polygons
# Create experiment plots}
\usage{
make_trial_plots_by_input(
  field,
  ab_lines_data,
  abline_type,
  plot_width,
  machine_width,
  harvester_width,
  section_num,
  headland_length,
  side_length,
  min_plot_length,
  max_plot_length,
  perpendicular,
  second_input = FALSE
)
}
\description{
# Create strips
# Shift the polygons
# Create experiment plots
}
