% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofpetrial-package.R
\docType{package}
\name{ofpetrial-package}
\alias{ofpetrial}
\alias{ofpetrial-package}
\title{ofpetrial: Design On-farm Field Trials}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Taro  Mieno \email{tmieno2@unl.edu}

Authors:
\itemize{
  \item Brittani Edge \email{bedge2@illinois.edu} [contributor]
}

}
\keyword{internal}
