% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_info}
\alias{plot_info}
\title{Plot information}
\format{
data.frame `plot_info`
A data frame with 1 rows and 10 columns:
\describe{
  \item{input_name}{input name}
  \item{unit_system}{measurement system (metric or imperial)}
  \item{machine_width}{width of the applicator/planter}
  \item{section_num}{number of the sections of the machine}
  \item{section_width}{width of a section of the machine}
  \item{harvester_width}{width of the harvester}
  \item{plot_width}{width of the plots to be made}
  \item{headland_length}{length of the headland}
  \item{side_length}{length of the side}
  \item{min_plot_length}{minimum plot length allowed}
  \item{max_plot_length}{maximum plot length allowed}
}
}
\usage{
plot_info
}
\description{
Plot information for creating experiment plots using `make_exp_plot()`. This data exists only for the purpose of making examples in some function references succinct.
}
\keyword{datasets}
