% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepre_plot_info.R
\name{prep_plot_fd}
\alias{prep_plot_fd}
\title{Prepare plot information for a two-input experiment (feet)}
\usage{
prep_plot_fd(
  form,
  machine_width,
  section_num,
  harvester_width,
  plot_width = c(NA, NA),
  headland_length = NA,
  side_length = NA,
  max_plot_width = 120,
  min_plot_length = 240,
  max_plot_length = 260
)
}
\arguments{
\item{form}{(vector of two elements)}

\item{machine_width}{(vector of two elements)}

\item{section_num}{(vector of two elements)}

\item{harvester_width}{(numeric)}

\item{plot_width}{(vector of two elements) Default is c(NA, NA).}

\item{headland_length}{(numeric) Default is NA.}

\item{side_length}{(numeric) Default is NA.}

\item{max_plot_width}{(numeric) Default is 120 feet.}

\item{min_plot_length}{(numeric) Default is 260 feet.}

\item{max_plot_length}{(numeric) Default is 240 feet.}
}
\value{
a tibble with plot information necessary to create experiment plots
}
\description{
Prepare plot information for a two-input experiment case. All the length values need to be specified in feet.
}
\examples{
form <- c("seed", "NH3")
machine_width <- c(12, 9)
section_num <- c(12, 1)
plot_width <- c(9.5, 36)
harvester_width <- 12
prep_plot_fd(form, machine_width, section_num, harvester_width, plot_width)

}
