% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepre_plot_info.R
\name{prep_plot_fs}
\alias{prep_plot_fs}
\title{Prepare plot information for a single-input experiment (feet)}
\usage{
prep_plot_fs(
  form,
  machine_width,
  section_num,
  harvester_width,
  plot_width = NA,
  headland_length = NA,
  side_length = NA,
  max_plot_width = 120,
  min_plot_length = 240,
  max_plot_length = 260
)
}
\arguments{
\item{form}{(character)}

\item{machine_width}{(numeric)}

\item{section_num}{(numeric)}

\item{harvester_width}{(numeric)}

\item{plot_width}{(numeric) Default is c(NA, NA).}

\item{headland_length}{(numeric) Default is NA.}

\item{side_length}{(numeric) Default is NA.}

\item{max_plot_width}{(numeric) Default is 120 feet.}

\item{min_plot_length}{(numeric) Default is 260 feet.}

\item{max_plot_length}{(numeric) Default is 240 feet.}
}
\value{
a tibble with plot information necessary to create experiment plots
}
\description{
Prepare plot information for a single-input experiment case. All the length values need to be specified in feet.
}
\examples{
form <- "seed"
machine_width <- 60
section_num <- 24
plot_width <- NA
harvester_width <- 30
prep_plot_fs(form, machine_width, section_num, harvester_width)

}
