% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepre_plot_info.R
\name{prep_plot_ms}
\alias{prep_plot_ms}
\title{Prepare plot information for a single-input experiment (meter)}
\usage{
prep_plot_ms(
  form,
  machine_width,
  section_num,
  harvester_width,
  plot_width = NA,
  headland_length = NA,
  side_length = NA,
  max_plot_width = measurements::conv_unit(120, "ft", "m"),
  min_plot_length = measurements::conv_unit(240, "ft", "m"),
  max_plot_length = measurements::conv_unit(260, "ft", "m")
)
}
\arguments{
\item{form}{(character)}

\item{machine_width}{(numeric)}

\item{section_num}{(numeric)}

\item{harvester_width}{(numeric)}

\item{plot_width}{(numeric) Default is c(NA, NA).}

\item{headland_length}{(numeric) Default is NA.}

\item{side_length}{(numeric) Default is NA.}

\item{max_plot_width}{(numeric) Default is 36.576 meter (120 feet).}

\item{min_plot_length}{(numeric) Default is 79.248 meter (260 feet).}

\item{max_plot_length}{(numeric) Default is 73.152 meter (240 feet).}
}
\value{
a tibble with plot information necessary to create experiment plots
}
\description{
Prepare plot information for a single-input experiment case. All the length values need to be specified in meter.
}
\examples{
form <- "seed"
machine_width <- 12
section_num <- 12
plot_width <- NA
harvester_width <- 24
prep_plot_ms(form, machine_width, section_num, harvester_width)

}
