% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rate_info}
\alias{rate_info}
\title{Rate information}
\format{
data.frame `rate_info`
A data frame with 1 rows and 7 columns:
\describe{
  \item{input_name}{input name}
  \item{design_type}{type of the trial design to be created}
  \item{gc_rate}{normal rate the grower would have used if not running an experiment}
  \item{unit}{unit of the input}
  \item{rates_data}{data.frame of rates and their ranks}
  \item{rank_seq_ws}{vector of the ranking of rates that will repeated within a strip}
  \item{rank_seq_as}{vector of the ranking of rates that will repeated as the first rate of the strips}
}
}
\usage{
rate_info
}
\description{
Rate information for assigning rates to the experiment plots using the `assign_rates()` function. This data exists only for the purpose of making examples in some function references succinct.
}
\keyword{datasets}
