% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_spatial.R
\name{st_extend_line}
\alias{st_extend_line}
\title{Extend a line}
\usage{
st_extend_line(line, multiplier)
}
\arguments{
\item{line}{sf # sf line object to be extended}

\item{multiplier}{scalar}
}
\value{
sfc of the extended line
}
\description{
Elongate a line by the user-specified multiplier
}
