% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_spatial.R
\name{st_shift}
\alias{st_shift}
\title{Shift an sf (spatial) object}
\usage{
st_shift(data_sf, shift, merge = TRUE)
}
\arguments{
\item{data_sf}{sf # the sf object to be shifted}

\item{shift}{vector # 2 by 1 vector with (x, y), where x is how much to shift in easting and y is how much to shift in northing}

\item{merge}{logical #  if true, then the shifted geometry replaces the existing geometry of the input sf (data_sf), otherwise only the geometry as sfc is returned}
}
\value{
Either sf or sfc depending on the merge parameter
}
\description{
Shift an sf object by user-specified amount in X and Y
}
