% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{td_single_input}
\alias{td_single_input}
\title{Trial design (single-input)}
\format{
tbl_df tbl data.frame `td_single_input`
A data frame with 1 rows and 9 columns:
\describe{
  \item{input_name}{input name}
  \item{input_type}{shorthand for the type of the input: "N" for nitrogen, "S" for seed, etc.}
  \item{trial_design}{experiment plots with input rats assigned as an sf object}
  \item{design_type}{type of the trial design used}
  \item{unit}{unit of the input}
  \item{abline_type}{(character) one of "free", "lock", "none" indicating the way ab-line is (or not) created}
  \item{ab_lines}{ab-lines for the applicator/planter as an sf object}
  \item{harvest_ab_lines}{ab-lines for the harvester as an sf object}
  \item{field_sf}{field boundary as an sf object}
  \item{harvest_width}{width of the harvester}
}
}
\usage{
td_single_input
}
\description{
Trial design data created by assigning rates to experiment plots running the `assign_rates()` function. This data exists only for the purpose of making examples in some function references succinct.
}
\keyword{datasets}
