% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{viz_ep}
\alias{viz_ep}
\title{Visualize experiment plots}
\usage{
viz_ep(exp_data, abline = TRUE)
}
\arguments{
\item{exp_data}{experiment plots created by make_exp_plots()}

\item{abline}{(logical) Default = TRUE. If TRUE, ab-lines are displayed along with the experiment plots}
}
\value{
ggplot2 figure
}
\description{
Visualize experiment plots (no rates assigned)
}
