% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{viz_exp_plots}
\alias{viz_exp_plots}
\title{Visualize experiment plots}
\usage{
viz_exp_plots(exp_data)
}
\arguments{
\item{exp_data}{experiment plots created by make_exp_plots()}
}
\value{
ggplot2 figure
}
\description{
Visualize experiment plots (no rates assigned)
}
