% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{viz_td}
\alias{viz_td}
\title{Visualize trial designs}
\usage{
viz_td(trial_design, abline = TRUE)
}
\arguments{
\item{trial_design}{trial design: experiment plots with input rates assigned}

\item{abline}{(logical) Default = TRUE. If TRUE, ab-lines are displayed along with the trial designs}
}
\value{
ggplot2 figure
}
\description{
Visualize trial designs plots (rates assigned)
}
