% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_rates.R
\name{add_blocks}
\alias{add_blocks}
\title{Add blocks to trial design}
\usage{
add_blocks(td)
}
\arguments{
\item{td}{trial design made by applying assign_rates() to experimental plots made by make_exp_plots()}
}
\value{
trial design with block_id added
}
\description{
Delineate blocks on a trial design and assign block id to all the plots
}
\examples{
#--- load rate information ---#
data(td_single_input)

#--- add blocks ---#
td_with_blocks <- add_blocks(td_single_input)

#--- take a look ---#
td_with_blocks$trial_design

#--- visualize ---#
viz(td_with_blocks, type = "block_id")

}
