% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_rates.R
\name{assign_rates}
\alias{assign_rates}
\title{Assign rates to the plots of experimental plots}
\usage{
assign_rates(exp_data, rate_info)
}
\arguments{
\item{exp_data}{experiment plots created by make_exp_plots()}

\item{rate_info}{rate information created by prep_rate()}
}
\value{
trial design as sf (experiment plots with rates assigned)
}
\description{
This functions assign input rates for the plots created by make_exp_plots() according to the rate designs specified by the user in rate_info, which can be created by prep_rateingle().
}
\examples{
#--- load experiment plots made by make_exp_plots() ---#
data(exp_data)
exp_data

#--- load rate information ---#
data(rate_info)
rate_info

#--- assign rates ---#
td <- assign_rates(exp_data, rate_info)

#--- visualization of the assigned rates ---#
viz(td)
}
