% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_rates_manually.R
\name{change_rate_by_block}
\alias{change_rate_by_block}
\title{Change the assigned rate by block}
\usage{
change_rate_by_block(td, input_name, block_ids, new_rates)
}
\arguments{
\item{td}{trial design}

\item{input_name}{(character) input name}

\item{block_ids}{(numeric) vector of block_ids}

\item{new_rates}{(numeric) vector of rates}
}
\value{
trial design with changed rates
}
\description{
Change the assigned rates by block on trial design
}
\examples{
#--- load rate information ---#
data(td_single_input)

#--- add blocks ---#
td_with_blocks <- add_blocks(td_single_input)

#--- change rates of some blocks ---#
block_ids <- c(3, 5)
new_rates <- c(180, 180)

td_modified <- change_rate_by_block(td_with_blocks, "NH3", block_ids, new_rates)

#--- visualize ---#
viz(td_modified)
}
