% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_rates.R
\name{change_rate_by_strip}
\alias{change_rate_by_strip}
\title{Change the assigned rate by strip}
\usage{
change_rate_by_strip(form, strip_ids, strip_rates, td)
}
\arguments{
\item{form}{(character) input name}

\item{strip_ids}{(numeric) vector of strip_ids}

\item{strip_rates}{(numeric) vector of rates}

\item{td}{trial design}
}
\value{
trial design with changed rates
}
\description{
Change the assigned rates by strip on trial design
}
\examples{
#--- load rate information ---#
data(trial_design)

#--- change rates of some strips ---#
strip_ids <- c(1, 6, 11, 16, 21)
strip_rates <- rep(0, length(strip_ids))

td_changed <- change_rate_by_strip("NH3", strip_ids, strip_rates, trial_design)

#--- visualize ---#
library(ggplot2)
ggplot(td_changed$trial_design[[1]]) +
  geom_sf(aes(fill = factor(rate))) +
  scale_fill_viridis_d() +
  theme_void()
}
