% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_rates.R
\name{make_input_rate_data}
\alias{make_input_rate_data}
\title{Create data of input rate information for a single input}
\usage{
make_input_rate_data(
  plot_info,
  gc_rate,
  unit,
  rates = NULL,
  min_rate = NULL,
  max_rate = NULL,
  num_rates = 5,
  design_type = NA
)
}
\arguments{
\item{plot_info}{(data.frame) plot information created by make_input_plot_data}

\item{gc_rate}{(numeric) Input ate the grower would have chosen if not running an experiment. This rate is assigned to the non-experiment part of the field. This rate also becomes one of the trial input rates unless you specify the trial rates directly using rates argument}

\item{unit}{(string) unit of input}

\item{rates}{(numeric vector) Default is NULL. Sequence of trial rates in the ascending order.}

\item{min_rate}{(numeric) minimum input rate. Ignored if rates are specified.}

\item{max_rate}{(numeric) maximum input rate. Ignored if rates are specified}

\item{num_rates}{(numeric) Default is 5. It has to be an even number if design_type is "ejca". Ignored if rates are specified.}

\item{design_type}{(string) type of trial design. available options are "jcl", "sparse", and "ejca". See for more details.}
}
\value{
data.frame of input rate information
}
\description{
Create data of input rate information for a single input. This can be used to assign rates to experimentl plots using assign_rates()
}
