% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trial_report.R
\name{make_trial_report}
\alias{make_trial_report}
\title{Create trial design report}
\usage{
make_trial_report(td, land_unit, units, trial_name, folder_path)
}
\arguments{
\item{td}{trial design created by assign_rates()}

\item{land_unit}{unit of land area in report ("acres" or "hectares")}

\item{units}{units for report ("imperial" or "metric")}

\item{trial_name}{(character) name of trial to be used in report}

\item{folder_path}{(character) path to the folder in which the report will be saved}
}
\value{
html document with trial design description and figures to guide trial implementation based on your machinery and plot sizes
}
\description{
This function creates a report describing the trial design created by the user with assign_rates() and includes figures showing machine alignment
}
\examples{
#--- load experiment made by assign_rates() ---#
data(td_single_input)
td_single_input

#--- create report ---#
td <- make_trial_report(td_single_input, "acres", "imperial", "Test Trial", folder_path)

# Rmd
# Wrapping up
}
