% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trial_report.R
\name{make_trial_report}
\alias{make_trial_report}
\title{Create trial design report}
\usage{
make_trial_report(td, trial_name = NA, folder_path = getwd())
}
\arguments{
\item{td}{trial design created by assign_rates()}

\item{trial_name}{(character) name of trial to be used in report}

\item{folder_path}{(character) path to the folder in which the report will be saved}
}
\value{
html document with trial design description and figures to guide trial implementation based on your machinery and plot sizes
}
\description{
This function creates a report describing the trial design created by the user with assign_rates() and includes figures showing machine alignment
}
\examples{
#--- load experiment made by assign_rates() ---#
data(td_single_input)
make_trial_report(
  td = td_single_input,
  folder_path = tempdir()
)
}
