% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofpetrial-package.R
\docType{package}
\name{ofpetrial-package}
\alias{ofpetrial}
\alias{ofpetrial-package}
\title{ofpetrial: Design on-Farm Precision Field Agronomic Trials}
\description{
A comprehensive system for designing and implementing on-farm precision field agronomic trials. You provide field data, tell 'ofpetrial' how to design a trial, and get readily-usable trial design files and a report checks the validity and reliability of the trial design.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://difm-brain.github.io/ofpetrial/}
  \item Report bugs at \url{https://github.com/DIFM-Brain/ofpetrial/issues}
}

}
\author{
\strong{Maintainer}: Taro Mieno \email{tmieno2@unl.edu} (\href{https://orcid.org/0000-0002-0614-0771}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Brittani Edge \email{bedge2@illinois.edu} (\href{https://orcid.org/0009-0005-2051-1048}{ORCID}) [contributor]
}

}
\keyword{internal}
