% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofpetrial-package.R
\docType{package}
\name{ofpetrial-package}
\alias{ofpetrial}
\alias{ofpetrial-package}
\title{ofpetrial: Design On-Farm Precision Field Agronomic Trials}
\description{
A comprehensive system for designing and implementing on-farm precision field agronomic trials. You provide field data, tell 'ofpetrial' how to design a trial, and get readily-usable trial design files and a report checks the validity and reliability of the trial design.
}
\author{
\strong{Maintainer}: Taro  Mieno \email{tmieno2@unl.edu} [copyright holder]

Authors:
\itemize{
  \item Brittani Edge \email{bedge2@illinois.edu} [contributor]
}

}
\keyword{internal}
