% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepre_plot_info.R
\name{prep_plot_fd}
\alias{prep_plot_fd}
\title{Prepare plot information for a two-input experiment (length in feet)}
\usage{
prep_plot_fd(
  input_name,
  machine_width,
  section_num,
  harvester_width,
  plot_width = c(NA, NA),
  headland_length = NA,
  side_length = NA,
  max_plot_width = 120,
  min_plot_length = 240,
  max_plot_length = 300
)
}
\arguments{
\item{input_name}{(character) A vector of two input names}

\item{machine_width}{(numeric) A vector of two numeric numbers in feet that indicate the width of the applicator or planter of the inputs}

\item{section_num}{(numeric) A vector of two numeric numbers that indicate the number of sections of the applicator or planter of the inputs}

\item{harvester_width}{(numeric) A numeric number in feet that indicates the width of the harvester}

\item{plot_width}{(numeric) A vector of two numeric numbers in feet that indicate the plot width for each of the two inputs. Default is c(NA, NA).}

\item{headland_length}{(numeric) A numeric number in feet that indicates the length of the headland (how long the non-experimental space is in the direction machines drive). Default is NA.}

\item{side_length}{(numeric) A numeric number in feet that indicates the length of the two sides of the field (how long the non-experimental space is in the direction perpendicular to the direction of machines). Default is NA.}

\item{max_plot_width}{(numeric) Maximum width of the plots in feet. Default is (36.576 meter).}

\item{min_plot_length}{(numeric) Minimum length of the plots in feet. Default is 240 feet (73.152 meter).}

\item{max_plot_length}{(numeric) Maximum length of the plots in feet. Default is 300 feet (91.44 meter).}
}
\value{
a tibble with plot information necessary to create experiment plots
}
\description{
Prepare plot information for a two-input experiment case. All the length values need to be specified in feet.
}
\examples{
input_name <- c("seed", "NH3")
machine_width <- c(12, 9)
section_num <- c(12, 1)
plot_width <- c(12, 36)
harvester_width <- 12
prep_plot_fd(input_name, machine_width, section_num, harvester_width, plot_width)

}
