% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{viz_td}
\alias{viz_td}
\title{Visualize trial designs}
\usage{
viz_td(trial_design)
}
\arguments{
\item{trial_design}{trial design: experiment plots with input rates assigned}
}
\description{
Visualize trial designs plots (rates assigned)
}
